[core] Add dry_run parameters to purge_files Procedure and display the list of deleted directories (#5342)

+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/procedure/PurgeFilesProcedure.java
@@ -21,6 +21,7 @@ package org.apache.paimon.flink.procedure;
 import org.apache.paimon.catalog.Catalog;
 import org.apache.paimon.catalog.Identifier;
 import org.apache.paimon.fs.FileIO;
+import org.apache.paimon.fs.FileStatus;
 import org.apache.paimon.fs.Path;
 import org.apache.paimon.table.FileStoreTable;
 import org.apache.paimon.table.Table;
@@ -29,8 +30,10 @@ import org.apache.flink.table.annotation.ArgumentHint;
 import org.apache.flink.table.annotation.DataTypeHint;
 import org.apache.flink.table.annotation.ProcedureHint;
 import org.apache.flink.table.procedure.ProcedureContext;
+import org.apache.flink.types.Row;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
 
 /**
@@ -45,20 +48,31 @@ public class PurgeFilesProcedure extends ProcedureBase {
 
     public static final String IDENTIFIER = "purge_files";
 
-    @ProcedureHint(argument = {@ArgumentHint(name = "table", type = @DataTypeHint("STRING"))})
-    public String[] call(ProcedureContext procedureContext, String tableId)
+    @ProcedureHint(
+            argument = {
+                @ArgumentHint(name = "table", type = @DataTypeHint("STRING")),
+                @ArgumentHint(name = "dry_run", type = @DataTypeHint("BOOLEAN"), isOptional = true)
+            })
+    public @DataTypeHint("ROW<purged_file_path STRING>") Row[] call(
+            ProcedureContext procedureContext, String tableId, Boolean dryRun)
             throws Catalog.TableNotExistException {
         Table table = catalog.getTable(Identifier.fromString(tableId));
         FileStoreTable fileStoreTable = (FileStoreTable) table;
         FileIO fileIO = fileStoreTable.fileIO();
         Path tablePath = fileStoreTable.snapshotManager().tablePath();
+        ArrayList<String> deleteDir;
         try {
-            Arrays.stream(fileIO.listStatus(tablePath))
+            FileStatus[] fileStatuses = fileIO.listStatus(tablePath);
+            deleteDir = new ArrayList<>(fileStatuses.length);
+            Arrays.stream(fileStatuses)
                     .filter(f -> !f.getPath().getName().contains("schema"))
                     .forEach(
                             fileStatus -> {
                                 try {
-                                    fileIO.delete(fileStatus.getPath(), true);
+                                    deleteDir.add(fileStatus.getPath().getName());
+                                    if (dryRun == null || !dryRun) {
+                                        fileIO.delete(fileStatus.getPath(), true);
+                                    }
                                 } catch (IOException e) {
                                     throw new RuntimeException(e);
                                 }
@@ -66,9 +80,10 @@ public class PurgeFilesProcedure extends ProcedureBase {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-        return new String[] {
-            String.format("Success purge files with table: %s.", fileStoreTable.name())
-        };
+
+        return deleteDir.isEmpty()
+                ? new Row[] {Row.of("There are no dir to be deleted.")}
+                : deleteDir.stream().map(Row::of).toArray(Row[]::new);
     }
 
     @Override
